+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/cache/AlarmTypeIndicatorCacheImpl.java
@@ -1,0 +1,91 @@
+package com.dwarfeng.bitalarm.impl.cache;
+
+import com.dwarfeng.bitalarm.sdk.bean.entity.FastJsonAlarmTypeIndicator;
+import com.dwarfeng.bitalarm.stack.bean.entity.AlarmTypeIndicator;
+import com.dwarfeng.bitalarm.stack.cache.AlarmTypeIndicatorCache;
+import com.dwarfeng.subgrade.impl.cache.RedisBatchBaseCache;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.stack.bean.key.ByteIdKey;
+import com.dwarfeng.subgrade.stack.exception.CacheException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+import java.util.List;
+
+@Repository
+public class AlarmTypeIndicatorCacheImpl implements AlarmTypeIndicatorCache {
+
+    @Autowired
+    private RedisBatchBaseCache<ByteIdKey, AlarmTypeIndicator, FastJsonAlarmTypeIndicator> batchBaseCache;
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public boolean exists(ByteIdKey key) throws CacheException {
+        return batchBaseCache.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public AlarmTypeIndicator get(ByteIdKey key) throws CacheException {
+        return batchBaseCache.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void push(AlarmTypeIndicator value, long timeout) throws CacheException {
+        batchBaseCache.push(value, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void delete(ByteIdKey key) throws CacheException {
+        batchBaseCache.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void clear() throws CacheException {
+        batchBaseCache.clear();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public boolean allExists(List<ByteIdKey> keys) throws CacheException {
+        return batchBaseCache.allExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public boolean nonExists(List<ByteIdKey> keys) throws CacheException {
+        return batchBaseCache.nonExists(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public List<AlarmTypeIndicator> batchGet(List<ByteIdKey> keys) throws CacheException {
+        return batchBaseCache.batchGet(keys);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void batchPush(List<AlarmTypeIndicator> entities, long timeout) throws CacheException {
+        batchBaseCache.batchPush(entities, timeout);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void batchDelete(List<ByteIdKey> keys) throws CacheException {
+        batchBaseCache.batchDelete(keys);
+    }
+}