+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/dao/AlarmTypeIndicatorDaoImpl.java
@@ -1,15 +1,14 @@
 package com.dwarfeng.bitalarm.impl.dao;
 
-import com.dwarfeng.bitalarm.impl.bean.entity.HibernateAlarmHistory;
-import com.dwarfeng.bitalarm.stack.bean.entity.AlarmHistory;
-import com.dwarfeng.bitalarm.stack.dao.AlarmHistoryDao;
+import com.dwarfeng.bitalarm.impl.bean.entity.HibernateAlarmTypeIndicator;
+import com.dwarfeng.bitalarm.stack.bean.entity.AlarmTypeIndicator;
+import com.dwarfeng.bitalarm.stack.dao.AlarmTypeIndicatorDao;
 import com.dwarfeng.subgrade.impl.dao.HibernateBatchBaseDao;
 import com.dwarfeng.subgrade.impl.dao.HibernateEntireLookupDao;
-import com.dwarfeng.subgrade.impl.dao.HibernatePresetLookupDao;
-import com.dwarfeng.subgrade.sdk.bean.key.HibernateLongIdKey;
+import com.dwarfeng.subgrade.sdk.bean.key.HibernateByteIdKey;
 import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
 import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
-import com.dwarfeng.subgrade.stack.bean.key.LongIdKey;
+import com.dwarfeng.subgrade.stack.bean.key.ByteIdKey;
 import com.dwarfeng.subgrade.stack.exception.DaoException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
@@ -18,103 +17,101 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 @Repository
-public class AlarmHistoryDaoImpl implements AlarmHistoryDao {
+public class AlarmTypeIndicatorDaoImpl implements AlarmTypeIndicatorDao {
 
     @Autowired
-    private HibernateBatchBaseDao<LongIdKey, HibernateLongIdKey, AlarmHistory, HibernateAlarmHistory> batchBaseDao;
+    private HibernateBatchBaseDao<ByteIdKey, HibernateByteIdKey, AlarmTypeIndicator, HibernateAlarmTypeIndicator> batchBaseDao;
     @Autowired
-    private HibernateEntireLookupDao<AlarmHistory, HibernateAlarmHistory> entireLookupDao;
-    @Autowired
-    private HibernatePresetLookupDao<AlarmHistory, HibernateAlarmHistory> presetLookupDao;
+    private HibernateEntireLookupDao<AlarmTypeIndicator, HibernateAlarmTypeIndicator> entireLookupDao;
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
-    public LongIdKey insert(AlarmHistory element) throws DaoException {
+    public ByteIdKey insert(AlarmTypeIndicator element) throws DaoException {
         return batchBaseDao.insert(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
-    public void update(AlarmHistory element) throws DaoException {
+    public void update(AlarmTypeIndicator element) throws DaoException {
         batchBaseDao.update(element);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
-    public void delete(LongIdKey key) throws DaoException {
+    public void delete(ByteIdKey key) throws DaoException {
         batchBaseDao.delete(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public boolean exists(LongIdKey key) throws DaoException {
+    public boolean exists(ByteIdKey key) throws DaoException {
         return batchBaseDao.exists(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public AlarmHistory get(LongIdKey key) throws DaoException {
+    public AlarmTypeIndicator get(ByteIdKey key) throws DaoException {
         return batchBaseDao.get(key);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
-    public List<LongIdKey> batchInsert(List<AlarmHistory> elements) throws DaoException {
+    public List<ByteIdKey> batchInsert(List<AlarmTypeIndicator> elements) throws DaoException {
         return batchBaseDao.batchInsert(elements);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
-    public void batchUpdate(List<AlarmHistory> elements) throws DaoException {
+    public void batchUpdate(List<AlarmTypeIndicator> elements) throws DaoException {
         batchBaseDao.batchUpdate(elements);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager")
-    public void batchDelete(List<LongIdKey> keys) throws DaoException {
+    public void batchDelete(List<ByteIdKey> keys) throws DaoException {
         batchBaseDao.batchDelete(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public boolean allExists(List<LongIdKey> keys) throws DaoException {
+    public boolean allExists(List<ByteIdKey> keys) throws DaoException {
         return batchBaseDao.allExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public boolean nonExists(List<LongIdKey> keys) throws DaoException {
+    public boolean nonExists(List<ByteIdKey> keys) throws DaoException {
         return batchBaseDao.nonExists(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public List<AlarmHistory> batchGet(List<LongIdKey> keys) throws DaoException {
+    public List<AlarmTypeIndicator> batchGet(List<ByteIdKey> keys) throws DaoException {
         return batchBaseDao.batchGet(keys);
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public List<AlarmHistory> lookup() throws DaoException {
+    public List<AlarmTypeIndicator> lookup() throws DaoException {
         return entireLookupDao.lookup();
     }
 
     @Override
     @BehaviorAnalyse
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public List<AlarmHistory> lookup(PagingInfo pagingInfo) throws DaoException {
+    public List<AlarmTypeIndicator> lookup(PagingInfo pagingInfo) throws DaoException {
         return entireLookupDao.lookup(pagingInfo);
     }
 
@@ -123,26 +120,5 @@ public class AlarmHistoryDaoImpl implements AlarmHistoryDao {
     @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
     public int lookupCount() throws DaoException {
         return entireLookupDao.lookupCount();
-    }
-
-    @Override
-    @BehaviorAnalyse
-    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public List<AlarmHistory> lookup(String preset, Object[] objs) throws DaoException {
-        return presetLookupDao.lookup(preset, objs);
-    }
-
-    @Override
-    @BehaviorAnalyse
-    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public List<AlarmHistory> lookup(String preset, Object[] objs, PagingInfo pagingInfo) throws DaoException {
-        return presetLookupDao.lookup(preset, objs, pagingInfo);
-    }
-
-    @Override
-    @BehaviorAnalyse
-    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
-    public int lookupCount(String preset, Object[] objs) throws DaoException {
-        return presetLookupDao.lookupCount(preset, objs);
     }
 }