+++ bitalarm-impl/src/main/java/com/dwarfeng/bitalarm/impl/service/AlarmTypeIndicatorMaintainServiceImpl.java
@@ -1,0 +1,107 @@
+package com.dwarfeng.bitalarm.impl.service;
+
+import com.dwarfeng.bitalarm.stack.bean.entity.AlarmTypeIndicator;
+import com.dwarfeng.bitalarm.stack.service.AlarmTypeIndicatorMaintainService;
+import com.dwarfeng.subgrade.impl.service.DaoOnlyEntireLookupService;
+import com.dwarfeng.subgrade.impl.service.GeneralBatchCrudService;
+import com.dwarfeng.subgrade.sdk.interceptor.analyse.BehaviorAnalyse;
+import com.dwarfeng.subgrade.stack.bean.dto.PagedData;
+import com.dwarfeng.subgrade.stack.bean.dto.PagingInfo;
+import com.dwarfeng.subgrade.stack.bean.key.ByteIdKey;
+import com.dwarfeng.subgrade.stack.exception.ServiceException;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+public class AlarmTypeIndicatorMaintainServiceImpl implements AlarmTypeIndicatorMaintainService {
+
+    @Autowired
+    private GeneralBatchCrudService<ByteIdKey, AlarmTypeIndicator> crudService;
+    @Autowired
+    private DaoOnlyEntireLookupService<AlarmTypeIndicator> entireLookupService;
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public boolean exists(ByteIdKey key) throws ServiceException {
+        return crudService.exists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public AlarmTypeIndicator get(ByteIdKey key) throws ServiceException {
+        return crudService.get(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public ByteIdKey insert(AlarmTypeIndicator element) throws ServiceException {
+        return crudService.insert(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void update(AlarmTypeIndicator element) throws ServiceException {
+        crudService.update(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void delete(ByteIdKey key) throws ServiceException {
+        crudService.delete(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager", readOnly = true)
+    public AlarmTypeIndicator getIfExists(ByteIdKey key) throws ServiceException {
+        return crudService.getIfExists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public ByteIdKey insertIfNotExists(AlarmTypeIndicator element) throws ServiceException {
+        return crudService.insertIfNotExists(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void updateIfExists(AlarmTypeIndicator element) throws ServiceException {
+        crudService.updateIfExists(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public void deleteIfExists(ByteIdKey key) throws ServiceException {
+        crudService.deleteIfExists(key);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public ByteIdKey insertOrUpdate(AlarmTypeIndicator element) throws ServiceException {
+        return crudService.insertOrUpdate(element);
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public PagedData<AlarmTypeIndicator> lookup() throws ServiceException {
+        return entireLookupService.lookup();
+    }
+
+    @Override
+    @BehaviorAnalyse
+    @Transactional(transactionManager = "hibernateTransactionManager")
+    public PagedData<AlarmTypeIndicator> lookup(PagingInfo pagingInfo) throws ServiceException {
+        return entireLookupService.lookup(pagingInfo);
+    }
+}