+++ bitalarm-node/bitalarm-node-all/src/main/java/com/dwarfeng/bitalarm/node/all/configuration/FastJsonConfiguration.java
@@ -1,10 +1,7 @@
 package com.dwarfeng.bitalarm.node.all.configuration;
 
 import com.alibaba.fastjson.parser.ParserConfig;
-import com.dwarfeng.bitalarm.sdk.bean.entity.FastJsonAlarmHistory;
-import com.dwarfeng.bitalarm.sdk.bean.entity.FastJsonAlarmInfo;
-import com.dwarfeng.bitalarm.sdk.bean.entity.FastJsonAlarmSetting;
-import com.dwarfeng.bitalarm.sdk.bean.entity.FastJsonCurrentAlarm;
+import com.dwarfeng.bitalarm.sdk.bean.entity.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.context.annotation.Configuration;
@@ -21,6 +18,7 @@ public class FastJsonConfiguration {
         ParserConfig.getGlobalInstance().addAccept(FastJsonAlarmHistory.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonCurrentAlarm.class.getCanonicalName());
         ParserConfig.getGlobalInstance().addAccept(FastJsonAlarmInfo.class.getCanonicalName());
+        ParserConfig.getGlobalInstance().addAccept(FastJsonAlarmTypeIndicator.class.getCanonicalName());
         LOGGER.debug("FastJson autotype 白名单配置完毕");
     }
 }