+++ bitalarm-stack/src/main/java/com/dwarfeng/bitalarm/stack/dao/AlarmInfoDao.java
@@ -6,7 +6,7 @@ import com.dwarfeng.subgrade.stack.dao.BatchBaseDao;
 import com.dwarfeng.subgrade.stack.dao.EntireLookupDao;
 
 /**
- * 当前报警数据访问层。
+ * 报警信息数据访问层。
  *
  * @author DwArFeng
  * @since 1.0.0